requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
opencv-python
numpy
pillow
transformers
albumentations
scikit-image
streamlit